+++ lib/signal.dart
@@ -1,10 +1,3 @@
 library signal;
 
-export 'src/AncestorChannelBuilder.dart';
-export 'src/AncestorChannelProvider.dart';
-export 'src/AvailableChannelBuilder.dart';
-export 'src/BaseState.dart';
-export 'src/ChannelSignal.dart';
-export 'src/OwnChannelBuilder.dart';
-export 'src/StateChannel.dart';
-export 'src/BaseLifeCycle.dart';
+export 'src/ex.dart';