+++ lib/src/BaseLifeCycle.dart
@@ -1,5 +1,0 @@
-abstract class BaseLifeCycle {
-  void initState();
-  void afterInitState();
-  void dispose();
-}