+++ lib/src/ex.dart
@@ -1,0 +1,5 @@
+export 'channelbuilder.dart';
+export 'channelprovider.dart';
+export 'channelsignal.dart';
+export 'statechannel.dart';
+export 'package:basestate/basestate.dart';